+++ lib/src/controllers/navigators/navigator_controller.dart
@@ -1,13 +1,13 @@
 // ignore_for_file: non_constant_identifier_names
-import '../../utils/platform/platform_stub.dart'
-    if (dart.library.io) '../../utils/platform/mobile_platform.dart'
-    if (dart.library.html) '../../utils/platform/web_platform.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_micro_app/flutter_micro_app.dart';
 import 'package:flutter_micro_app/src/presentation/pages/page_transition/micro_page_transition.dart';
 import 'package:flutter_micro_app/src/utils/enums/navigator_status.dart';
 
+import '../../utils/platform/platform_stub.dart'
+    if (dart.library.io) '../../utils/platform/mobile_platform.dart'
+    if (dart.library.html) '../../utils/platform/web_platform.dart';
 import 'navigator_event_controller.dart';
 
 /// instance hidden
@@ -22,12 +22,21 @@ MicroAppNavigatorController get NavigatorInstance {
 
 /// [MicroAppNavigatorController]
 class MicroAppNavigatorController extends RouteObserver<PageRoute<dynamic>> {
-  Map<String, PageBuilder> pageBuilders = {};
+  final Map<String, PageBuilder> _pageBuilders = {};
   final MicroAppNavigatorEventController eventController;
 
   MicroAppNavigatorController(this.eventController) {
     _navigatorInstance = this;
   }
+
+  /// [addPageBuilders]
+  void addPageBuilders(Map<String, PageBuilder> map) {
+    _pageBuilders.addAll(map);
+  }
+
+  /// [hasRoute]
+  bool hasRoute(String route) => _pageBuilders.containsKey(route);
+  Map<String, PageBuilder> get pageBuilders => _pageBuilders;
 
   /// [navigatorKey]
   GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
@@ -38,7 +47,7 @@ class MicroAppNavigatorController extends RouteObserver<PageRoute<dynamic>> {
   /// [getPageWidget]
   Widget getPageWidget(String route, BuildContext context,
           {Object? arguments, String? type, Widget? orElse}) =>
-      getPageBuilder(route)?.call(context, arguments) ??
+      getPageBuilder(route)?.builder.call(context, arguments) ??
       orElse ??
       const SizedBox.shrink();
 
@@ -177,19 +186,20 @@ class MicroAppNavigatorController extends RouteObserver<PageRoute<dynamic>> {
       return null;
     }
 
-    final transitionType = MicroAppPreferences.config.pageTransitionType;
+    final transitionType = pageBuilder.transitionType ??
+        MicroAppPreferences.config.pageTransitionType;
 
     if (transitionType != MicroPageTransitionType.platform) {
       return MicroPageTransition(
-          pageBuilder: pageBuilder, type: transitionType);
+          pageBuilder: pageBuilder.builder, type: transitionType);
     } else {
       if (isIOS) {
         return CupertinoPageRoute(
-          builder: (context) => pageBuilder(context, routeArguments),
+          builder: (context) => pageBuilder.builder(context, routeArguments),
         );
       } else {
         return MaterialPageRoute(
-          builder: (context) => pageBuilder(context, routeArguments),
+          builder: (context) => pageBuilder.builder(context, routeArguments),
         );
       }
     }
@@ -198,7 +208,7 @@ class MicroAppNavigatorController extends RouteObserver<PageRoute<dynamic>> {
   PageBuilder? getPageBuilder(String? name, {MicroAppBaseRoute? baseRoute}) {
     PageBuilder? pageBuilder;
     if (baseRoute == null || baseRoute.toString().isEmpty) {
-      pageBuilder = pageBuilders[name];
+      pageBuilder = _pageBuilders[name];
     } else {
       final buildersForBaseRoute = filterPageBuilderForBaseRoute(baseRoute);
       pageBuilder = buildersForBaseRoute[name];
@@ -208,7 +218,7 @@ class MicroAppNavigatorController extends RouteObserver<PageRoute<dynamic>> {
 
   Map<String, PageBuilder> filterPageBuilderForBaseRoute(
       MicroAppBaseRoute baseRoute) {
-    var filteredMap = Map<String, PageBuilder>.from(pageBuilders);
+    var filteredMap = Map<String, PageBuilder>.from(_pageBuilders);
     filteredMap
         .removeWhere((key, value) => !key.startsWith(baseRoute.toString()));
     return filteredMap;